Remove AppEngineFactory interface
