+++ src/main/java/com/google/cloud/tools/maven/CloudSdkAppEngineFactory.java
@@ -39,11 +39,31 @@ import java.nio.file.Path;
 import org.apache.maven.plugin.logging.Log;
 
 /** Factory for App Engine dependencies. */
-public class CloudSdkAppEngineFactory implements AppEngineFactory {
+public class CloudSdkAppEngineFactory {
 
   private CloudSdkFactory cloudSdkFactory;
   private CloudSdkMojo mojo;
   private CloudSdkOperationsFactory cloudSdkOperationsFactory;
+
+  /** Supported dev app server versions. */
+  public enum SupportedDevServerVersion {
+    V1,
+    V2ALPHA;
+
+    /**
+     * Parses {@code versionString} into a {@link SupportedDevServerVersion}. The aim is to let the
+     * users use lowercase in version strings.
+     */
+    public static SupportedDevServerVersion parse(String versionString) {
+      if ("1".equals(versionString)) {
+        return V1;
+      } else if ("2-alpha".equals(versionString)) {
+        return V2ALPHA;
+      } else {
+        throw new IllegalArgumentException("Unsupported version value: " + versionString);
+      }
+    }
+  }
 
   /**
    * Constructs a new CloudSdkAppEngineFactory
@@ -63,7 +83,11 @@ public class CloudSdkAppEngineFactory implements AppEngineFactory {
     this.cloudSdkOperationsFactory = cloudSdkOperationsFactory;
   }
 
-  @Override
+  /**
+   * Constructs an object used for standard staging
+   *
+   * @return a cloud sdk object for standard staging
+   */
   public AppEngineStandardStaging standardStaging() {
     try {
       return cloudSdkFactory.standardStaging(defaultCloudSdkBuilder().build());
@@ -72,12 +96,20 @@ public class CloudSdkAppEngineFactory implements AppEngineFactory {
     }
   }
 
-  @Override
+  /**
+   * Constructs an object used for flexible staging
+   *
+   * @return a cloud sdk object for flexible staging
+   */
   public AppEngineFlexibleStaging flexibleStaging() {
     return cloudSdkFactory.flexibleStaging();
   }
 
-  @Override
+  /**
+   * Constructs an object used for deployment
+   *
+   * @return a cloud sdk object for deployment
+   */
   public AppEngineDeployment deployment() {
     try {
       return cloudSdkFactory.deployment(defaultCloudSdkBuilder().build());
@@ -86,7 +118,11 @@ public class CloudSdkAppEngineFactory implements AppEngineFactory {
     }
   }
 
-  @Override
+  /**
+   * Constructs a dev server for the run goal
+   *
+   * @return a dev server for the run goal
+   */
   public AppEngineDevServer devServerRunSync(SupportedDevServerVersion version) {
     return createDevServerForVersion(version);
   }
@@ -111,7 +147,11 @@ public class CloudSdkAppEngineFactory implements AppEngineFactory {
     }
   }
 
-  @Override
+  /**
+   * Constructs a dev server
+   *
+   * @return a dev server for the runAsync goal
+   */
   public AppEngineDevServer devServerRunAsync(
       int startSuccessTimeout, SupportedDevServerVersion version) {
     CloudSdk.Builder builder =
@@ -123,12 +163,20 @@ public class CloudSdkAppEngineFactory implements AppEngineFactory {
     }
   }
 
-  @Override
+  /**
+   * Constructs a dev server for the stop goal
+   *
+   * @return a dev server for the stop goal
+   */
   public AppEngineDevServer devServerStop(SupportedDevServerVersion version) {
     return createDevServerForVersion(version);
   }
 
-  @Override
+  /**
+   * Constructs a cloud sdk object used for the genRepoInfoFile goal
+   *
+   * @return a cloud sdk object used for the genRepoInfoFile goal
+   */
   public GenRepoInfoFile genRepoInfoFile() {
     try {
       return cloudSdkFactory.genRepoInfoFile(defaultCloudSdkBuilder().build());
