+++ src/test/java/com/google/cloud/tools/maven/StopMojoTest.java
@@ -22,7 +22,7 @@ import static org.mockito.Mockito.when;
 
 import com.google.cloud.tools.appengine.api.AppEngineException;
 import com.google.cloud.tools.appengine.api.devserver.AppEngineDevServer;
-import com.google.cloud.tools.maven.AppEngineFactory.SupportedDevServerVersion;
+import com.google.cloud.tools.maven.CloudSdkAppEngineFactory.SupportedDevServerVersion;
 import java.io.IOException;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
